fundvision_backend/requirements.txt
# Core Django
Django==5.0.6
djangorestframework==3.15.2

# Database
PyMySQL==1.1.1
cryptography==42.0.8

# Authentication
djangorestframework-simplejwt==5.3.1

# CORS
django-cors-headers==4.4.0

# WebSockets / Channels
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2

# Task Queue
celery==5.4.0
redis==5.0.7
django-celery-beat==2.6.0

# API Documentation
drf-spectacular==0.27.2

# Rate Limiting
django-ratelimit==4.1.0

# HTTP Requests (for external market data APIs)
httpx==0.27.0
requests==2.32.3

# Environment Variables
python-decouple==3.8
python-dotenv==1.0.1

# Image Handling (profile pictures)
Pillow>=11.0.0

# Utilities
django-extensions==3.2.3
django-filter==24.2
